categories-page: display categories and posts in hierarchical structure
In Hexo, categories are designed to be hierarchical. However in our
theme the categories page is not able to display such hierarchy because
it simply loops over site.categories, which is a plain list of all
categories.

For example, if a post P is under category A, then subcategory B then
subcategory C, then in categories page there will three categories
A, B, C, each with a P under them.

With hierarchicalListHelper, P then will only be displayed under
category A > subcategory B > subcategory C. Note category A and
subcategory B will not be shown at all if there are no other posts
directly under them.